\name{addAndDownloadData}
\alias{addAndDownloadData}
\title{Add a Pre-Existing Object to the Package}
\arguments{
  \item{x}{: DataPackage}

  \item{identifier}{: character - the identifier of the
  object to act upon}

  \item{...}{(not yet used)}
}
\description{
  addAndDownloadData downloads a D1Object to the
  DataPackage, using the provided identifier string to
  retrieve from the DataONE system.
}
\author{
  rnahf
}

