\name{addData}
\alias{addData}
\title{Add a D1Object to the DataPackage}
\arguments{
  \item{x}{: DataPackage}

  \item{d1object}{: D1Object}

  \item{...}{: (not yet used)}
}
\description{
  Includes the D1Object in the DataPackage data Map, making
  it available for retrieval and eventual upload (via
  createPackage)
}
\author{
  rnahf
}

