% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_request.R
\name{auth_post}
\alias{auth_post}
\title{POST a resource with authenticated credentials.}
\usage{
auth_post(url, encode = "multipart", body = NULL, node)
}
\arguments{
\item{url}{The URL to be accessed via authenticated POST}

\item{encode}{the type of encoding to use for the POST body, defaults to 'multipart'}

\item{body}{a list of data to be included in the body of the POST request}

\item{node}{The D1Node object that the request will be made to.}
}
\value{
the HTTP response from the request
}
\description{
POST data to a URL using an HTTP POST request using authentication credentials 
provided in a client certificate.  Authenticated access depends on the suggested
openssl package. If the openssl package is not installed, then the request fails.
}
