% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\docType{methods}
\name{AuthenticationManager}
\alias{AuthenticationManager}
\alias{AuthenticationManager,ANY-method}
\title{Create an AuthenticationManager object}
\usage{
AuthenticationManager(...)

\S4method{AuthenticationManager}{ANY}()
}
\arguments{
\item{...}{(Not yet used)}
}
\value{
the AuthenticationManager object
}
\description{
Construct an instance of AuthenticationManager to provide mechanisms to load, verify, and 
display DataONE authentication information.
}
