% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNode.R
\name{CNode}
\alias{CNode}
\alias{CNode,ANY-method}
\alias{CNode,character-method}
\title{Create a CNode object.}
\usage{
CNode(x, ...)

\S4method{CNode}{ANY}()

\S4method{CNode}{character}(x)
}
\arguments{
\item{x}{The label for the DataONE environment to be using ('PROD','STAGING', 'STAGING2,'SANDBOX', 'SANDBOX2','DEV', 'DEV2')}

\item{...}{(not yet used)}
}
\value{
the CNode object representing the DataONE environment
}
\description{
Create a CNode object.
}
\details{
For an explanation of DataONE Coordinating Nodes, see the 
section \emph{"DataONE Environments"} in the overview vignette by entering the R command: \code{vignette("dataone-overview")}.
}
\examples{
\dontrun{
cn <- CNode("PROD")
}
}
\seealso{
\code{\link[=CNode-class]{CNode}}{ class description.}
}
