\name{convert.csv}
\alias{convert.csv}
\title{Convert a DataFrame to Standard CSV}
\usage{
  convert.csv(df, ...)
}
\arguments{
  \item{df}{: the dataFrame}

  \item{...}{}
}
\value{
  the dataframe serialized as a .csv
}
\description{
  Convert a DataFrame to Standard CSV
}
\author{
  Matt Jones
}

