% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Node.R
\name{d1_errors}
\alias{d1_errors}
\title{This function parses a DataONE service response message for errors, and extracts and
prints error information.}
\usage{
d1_errors(x)
}
\arguments{
\item{x}{The DataONE service response}
}
\description{
This function parses a DataONE service response message for errors, and extracts and
prints error information.
}
