% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{methods}
\name{initialize,D1Client-method}
\alias{initialize,D1Client-method}
\alias{D1Client-initialize}
\title{Initialize a D1Client object}
\usage{
\S4method{initialize}{D1Client}(.Object, cn = NA, mn = NA,
  env = as.character(NA), mNodeid = as.character(NA))
}
\arguments{
\item{.Object}{A D1client object.}

\item{cn}{The Member Node object to associate this D1Client with.}

\item{mn}{The Member Node object to associate this D1Client with.}

\item{env}{The DataONE environment to initialize this D1Client with, e.g. "PROD", "STAGING", "SANDBOX", "DEV"}

\item{mNodeid}{The node identifier of the Member Node to associate with this D1Client.}
}
\description{
Initialize a D1Client object
}
\examples{
\dontrun{
library(dataone)
d1c <- D1Client("PROD", "urn:node:KNB")
}
}
\seealso{
\code{\link[=D1Client-class]{dataone}}{ class description.}
}
