% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Node.R
\docType{methods}
\name{initialize,D1Node-method}
\alias{D1Node-initialize}
\alias{initialize,D1Node-method}
\title{Initialize a D1Node}
\usage{
\S4method{initialize}{D1Node}(.Object)
}
\arguments{
\item{.Object}{the D1Node object}
}
\description{
Initialize a D1Node
}

