% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Node.R
\docType{methods}
\name{D1Node}
\alias{D1Node}
\alias{D1Node,XMLInternalElementNode-method}
\title{Create a D1Node object.}
\usage{
D1Node(xml, ...)

\S4method{D1Node}{XMLInternalElementNode}(xml)
}
\arguments{
\item{xml}{An XML object that describes the node to be initialized (see \link{listNodes}).}

\item{...}{(not yet used)}
}
\value{
the Node object representing the DataONE environment
}
\description{
Create a D1Node object.
}

