% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Object.R
\docType{class}
\name{D1Object-class}
\alias{D1Object-class}
\title{D1Object (Defunct) is a representation of a DataObject.}
\description{
D1Object has been defunct in favor of datapack::DataObject, which provides
a wrapper for data and associated SystemMetadata.
}
\section{Slots}{

\describe{
\item{\code{dataObject}}{A backing instance of a DataObject, to which all methods and state are proxied}
}}

\section{Methods}{

\describe{
\item{\code{\link[=D1Object-initialize]{D1Object-initialize()}}}{: Initialize a D1Object}
\item{\code{\link[=getData]{getData()}}}{: Get the data content of a specified D1Object.}
\item{\code{\link[=getIdentifier]{getIdentifier()}}}{: Get the identifier of the D1Object.}
\item{\code{\link[=getFormatId]{getFormatId()}}}{: Get the formatId of the D1Object}
\item{\code{\link[=setPublicAccess]{setPublicAccess()}}}{: Add a Rule to the AccessPolicy to make the object publicly readable.}
\item{\code{\link[=canRead]{canRead()}}}{: Test whether the provided subject can read an object.}
\item{\code{\link[=asDataFrame]{asDataFrame()}}}{: Return the D1Object as a data.frame.}
}
}

\seealso{
\code{\link[=dataone]{dataone()}} package description.
}
\keyword{classes}
