% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Object.R
\docType{methods}
\name{initialize,D1Object-method}
\alias{initialize,D1Object-method}
\alias{D1Object-initialize}
\title{Initialize a D1Object}
\usage{
\S4method{initialize}{D1Object}(.Object, id, data, format,
  mnNodeId = as.character(NA))
}
\arguments{
\item{.Object}{A D1Object instance.}

\item{id}{The identifier for the object}

\item{data}{An R object (data or metadata) that this D1Object contains.}

\item{format}{The Object format.}

\item{mnNodeId}{The DataONE node identifier associated with this object, i.e. "urn:node:KNB"}
}
\description{
Initialize a D1Object
}
\seealso{
\code{\link[=D1Object-class]{D1Object}}{ class description.}
}
