% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Object.R
\name{D1Object}
\alias{D1Object}
\title{Create a D1Object instance.}
\usage{
D1Object(...)
}
\arguments{
\item{...}{(additional arguments)}
}
\value{
the D1Object instance
}
\description{
Create a D1Object instance.
}
\seealso{
\code{\link[=D1Object-class]{D1Object}}{ class description.}
}
