% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractTableDescriber.R, R/EMLParser.R
\name{data.characterEncoding}
\alias{data.characterEncoding}
\alias{data.characterEncoding,EMLParser,numeric-method}
\title{CharacterEncoding}
\usage{
data.characterEncoding(x, index, ...)

\S4method{data.characterEncoding}{EMLParser,numeric}(x, index)
}
\arguments{
\item{x}{the TableDescriber}

\item{index}{index of the table within the document}

\item{...}{Additional parameters}
}
\value{
the encoding used when serializing the data
}
\description{
The character encoding used, for example "UTF-8"
}
\author{
rnahf
}
