% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractTableDescriber.R, R/EMLParser.R
\docType{methods}
\name{data.tableAttributeNames}
\alias{data.tableAttributeNames}
\alias{data.tableAttributeNames,EMLParser,numeric-method}
\title{returns the attribute names}
\usage{
data.tableAttributeNames(x, index, ...)

\S4method{data.tableAttributeNames}{EMLParser,numeric}(x, index)
}
\arguments{
\item{x}{- the TableDescriber instance}

\item{index}{- the index of the table to get results for}

\item{...}{(not yet used)}
}
\value{
the attribute (column) names of the data
}
\description{
THe attribute names are defined in the metadata document for 
the specified data table
}
\author{
rnahf
}
