% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractTableDescriber.R, R/EMLParser.R
\docType{methods}
\name{data.tableAttributeOrientation}
\alias{data.tableAttributeOrientation}
\alias{data.tableAttributeOrientation,EMLParser,numeric-method}
\title{The Attribute (Header) Orientation}
\usage{
data.tableAttributeOrientation(x, index, ...)

\S4method{data.tableAttributeOrientation}{EMLParser,numeric}(x, index)
}
\arguments{
\item{x}{- the TableDescriber}

\item{index}{- the index of the table within the document}

\item{...}{Additional parameters}
}
\value{
legal values  are "columns" | "rows"
}
\description{
Which way to the attribute headers run?  Most data has a header row
where the attribute names go across "columns", in row in which case, the 
return value for this method should be "columns."
}
\note{
this is the opposite question from how records are organized!!
}
\author{
rnahf
}
