% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractTableDescriber.R, R/EMLParser.R
\docType{methods}
\name{data.tableAttributeStorageTypes}
\alias{data.tableAttributeStorageTypes}
\alias{data.tableAttributeStorageTypes,EMLParser,numeric-method}
\title{returns the attributes' data storage types}
\usage{
data.tableAttributeStorageTypes(x, index, ...)

\S4method{data.tableAttributeStorageTypes}{EMLParser,numeric}(x, index)
}
\arguments{
\item{x}{- the TableDescriber instance}

\item{index}{- the index of the table to get results for}

\item{...}{(not yet used)}
}
\value{
the data storage types of the attributes
}
\description{
The attributes' data storage types are defined in the metadata document for 
the specified data table
}
\author{
rnahf
}
