% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractTableDescriber.R, R/EMLParser.R
\docType{methods}
\name{data.tableAttributeTypes}
\alias{data.tableAttributeTypes}
\alias{data.tableAttributeTypes,EMLParser,numeric-method}
\title{returns the attributes' data types}
\usage{
data.tableAttributeTypes(x, index, ...)

\S4method{data.tableAttributeTypes}{EMLParser,numeric}(x, index)
}
\arguments{
\item{x}{- the TableDescriber instance}

\item{index}{- the index of the table to get results for}

\item{...}{(not yet used)}
}
\value{
the data types of the attributes
}
\description{
The attributes' data types are defined in the metadata document for 
the specified data table
}
\author{
rnahf
}
