% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractTableDescriber.R, R/EMLParser.R
\name{data.tableFieldDelimiter}
\alias{data.tableFieldDelimiter}
\alias{data.tableFieldDelimiter,EMLParser,numeric-method}
\title{Field Delimiter}
\usage{
data.tableFieldDelimiter(x, index, ...)

\S4method{data.tableFieldDelimiter}{EMLParser,numeric}(x, index)
}
\arguments{
\item{x}{the TableDescriber}

\item{index}{index of the table within the document}

\item{...}{Additional parameters}
}
\value{
the field delimiter(s) of the data object being described
}
\description{
Get the table field delimiter.
}
\author{
rnahf
}
