% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractTableDescriber.R, R/EMLParser.R
\name{data.tableMissingValueCodes}
\alias{data.tableMissingValueCodes}
\alias{data.tableMissingValueCodes,EMLParser,numeric-method}
\title{returns missing value codes}
\usage{
data.tableMissingValueCodes(x, index, ...)

\S4method{data.tableMissingValueCodes}{EMLParser,numeric}(x, index)
}
\arguments{
\item{x}{- the TableDescriber instance}

\item{index}{- the index of the table to get results for}

\item{...}{(not yet used)}
}
\value{
vector of missing value codes
}
\description{
the missing value codes are defined in the metadata document for 
the specified data table
}
\author{
rnahf
}
