% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractTableDescriber.R, R/EMLParser.R
\docType{methods}
\name{data.tableQuoteCharacter}
\alias{data.tableQuoteCharacter}
\alias{data.tableQuoteCharacter,EMLParser,numeric-method}
\alias{data.tableQuoteCharacter}
\title{Quote Character}
\usage{
data.tableQuoteCharacter(x, index, ...)

\S4method{data.tableQuoteCharacter}{EMLParser,numeric}(x, index)
}
\arguments{
\item{x}{the TableDescriber}

\item{index}{index of the table within the document}

\item{...}{Additional parameters}
}
\value{
the quote characters(s) for the data object being described
}
\description{
Get the table quote character.
}
\author{
rnahf
}
