% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractTableDescriber.R, R/EMLParser.R
\docType{methods}
\name{data.tableSkipLinesHeader}
\alias{data.tableSkipLinesHeader}
\alias{data.tableSkipLinesHeader,EMLParser,numeric-method}
\title{Number of lines to skip before reading data}
\usage{
data.tableSkipLinesHeader(x, index, ...)

\S4method{data.tableSkipLinesHeader}{EMLParser,numeric}(x, index)
}
\arguments{
\item{x}{- the TableDescriber}

\item{index}{- the index of the table within the document}

\item{...}{Additional parameters}
}
\value{
the number of lines to skip
}
\description{
The specified number of lines are skipped.
}
\seealso{
\code{help(read.table)}
}
\author{
rnahf
}
