% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Node.R, R/CNode.R, R/MNode.R
\docType{methods}
\name{describe}
\alias{describe}
\alias{describe,CNode-method}
\alias{describe,MNode-method}
\title{Efficiently get systemmetadat for an object.}
\usage{
describe(x, ...)

\S4method{describe}{CNode}(x, pid)

\S4method{describe}{MNode}(x, pid)
}
\arguments{
\item{x}{The MNode instance from which the identifier will be generated}

\item{...}{(Not yet used)}

\item{pid}{Identifier for the object in question. May be either a PID or a SID. Transmitted as
part of the URL path and must be escaped accordingly.}
}
\value{
A list of header elements
}
\description{
This method provides a lighter weight mechanism than getSystemMetadata() for a client to
determine basic properties of the referenced object.
}
\examples{
\dontrun{
library(dataone)
mn_uri <- "https://knb.ecoinformatics.org/knb/d1/mn/v1"
mn <- MNode(mn_uri)
pid <- "knb.473.1"
describe(mn, pid)
describe(mn, "adfadf") # warning message when wrong pid
}
}
\seealso{
\url{http://mule1.dataone.org/ArchitectureDocs-current/apis/MN_APIs.html#MNRead.describe}
}

