% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractTableDescriber.R, R/EMLParser.R
\docType{methods}
\name{documented.d1Identifiers}
\alias{documented.d1Identifiers}
\alias{documented.d1Identifiers,EMLParser-method}
\alias{documented.d1Identifiers}
\title{Get DataONE identifiers}
\usage{
documented.d1Identifiers(x, ...)

\S4method{documented.d1Identifiers}{EMLParser}(x)
}
\arguments{
\item{x}{the TableDescriber}

\item{...}{Additional parameters}
}
\value{
vector of dataONE identifiers
}
\description{
Get the DataONE identifier associated with each table
}
\author{
rnahf
}
