% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractTableDescriber.R, R/EMLParser.R
\docType{methods}
\name{documented.entityNames}
\alias{documented.entityNames}
\alias{documented.entityNames,EMLParser-method}
\alias{documented.entityNames}
\title{Get the entity names associated with each table}
\usage{
documented.entityNames(x, ...)

\S4method{documented.entityNames}{EMLParser}(x)
}
\arguments{
\item{x}{the TableDescriber}

\item{...}{Additional parameters}
}
\value{
vector of entity names
}
\description{
The entity names associated with each table are returned.
}
\author{
rnahf
}
