% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNode.R
\name{echoCredentials}
\alias{echoCredentials}
\alias{echoCredentials,CNode-method}
\title{Echo the credentials used to make the call.}
\usage{
echoCredentials(x, ...)

\S4method{echoCredentials}{CNode}(x)
}
\arguments{
\item{x}{The coordinating node to send the request to.}

\item{...}{(Not yet used)}
}
\value{
A list containing authentication info.
}
\description{
This method can be used to verify the client certificate is valid 
and contains the expected information.
}
\details{
The authentication credentials contained in the X.509 certificate or
authentication token are send with the request.
}
\examples{
\dontrun{
cn <- CNode("STAGING")
creds <- echoCredentials(cn)
print(creds$person$subject)
}
}
