% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMLParser.R
\docType{methods}
\name{EMLParser}
\alias{EMLParser}
\alias{EMLParser,D1Object-method}
\title{Construct an EML parser object.}
\usage{
EMLParser(d1Object, ...)

\S4method{EMLParser}{D1Object}(d1Object)
}
\arguments{
\item{d1Object}{The D1Object to obtain data from.}

\item{...}{Additional parameters}
}
\description{
Construct an EML parser object.
}
