% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Node.R
\docType{methods}
\name{encodeSolr}
\alias{encodeSolr}
\alias{encodeSolr,character-method}
\title{Encode the input for Solr Queries}
\usage{
encodeSolr(x, ...)

\S4method{encodeSolr}{character}(x, ...)
}
\arguments{
\item{x}{: a string to encode}

\item{...}{(not yet used.)}
}
\value{
the encoded form of the input
}
\description{
Treating all special characters and spaces as literals, backslash escape special
characters, and double-quote if necessary.
}
\examples{
encodeSolr("this & that")
}
