% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_request.R
\name{get_user_agent}
\alias{get_user_agent}
\title{User agent string}
\usage{
get_user_agent()
}
\description{
Get a string representation of the user agent to be sent to the server along
with other request details.
}
