% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\docType{methods}
\name{getAuthMethod}
\alias{getAuthMethod}
\alias{getAuthMethod,AuthenticationManager-method}
\title{Get the current valid authentication mechanism.}
\usage{
getAuthMethod(.Object, ...)

\S4method{getAuthMethod}{AuthenticationManager}(.Object, node)
}
\arguments{
\item{.Object}{An AuthenticationManager instance}

\item{...}{(Not yet used)}

\item{node}{A D1Node instance to determine the authentication method for.}
}
\value{
The current authentication mechanism as a character string, either "token" or "cert".
}
\description{
Get the current valid authentication mechanism.
}

