% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\docType{methods}
\name{getAuthSubject}
\alias{getAuthSubject}
\alias{getAuthSubject,AuthenticationManager-method}
\title{Get the authentication subject.}
\usage{
getAuthSubject(.Object, ...)

\S4method{getAuthSubject}{AuthenticationManager}(.Object, node)
}
\arguments{
\item{.Object}{an AuthenticationManager instance}

\item{...}{(Not yet used)}

\item{node}{A D1Node instance}
}
\value{
the DataONE Subject that is your client's identity
}
\description{
Get the authentication subject.
}

