% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\docType{methods}
\name{getCert}
\alias{getCert}
\alias{getCert,AuthenticationManager-method}
\title{Get the DataONE X.509 Certificate location.}
\usage{
getCert(.Object, ...)

\S4method{getCert}{AuthenticationManager}(.Object)
}
\arguments{
\item{.Object}{an AuthenticationManager instance}

\item{...}{(Not yet used)}
}
\value{
The filename of the current DataONE X.509 Certificate if it is available.
}
\description{
Get the DataONE X.509 Certificate location.
}
