% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CertificateManager.R
\name{getCertExpires}
\alias{getCertExpires}
\alias{getCertExpires,CertificateManager-method}
\title{Show the date and time when an X.509 certificate expires.}
\usage{
getCertExpires(x, ...)

\S4method{getCertExpires}{CertificateManager}(x)
}
\arguments{
\item{x}{a CertificateManager instance}

\item{...}{(Not yet used)}
}
\value{
POSIXct value
}
\description{
Each X.509 has a range of certificate validity times.  This method returns the X.509 
\code{'notAfter'} field formatted as a \code{'POSIXct'} date value.
}
