% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\docType{methods}
\name{getCertInfo}
\alias{getCertInfo}
\alias{getCertInfo,AuthenticationManager-method}
\title{Get X.509 Certificate information}
\usage{
getCertInfo(.Object)

\S4method{getCertInfo}{AuthenticationManager}(.Object)
}
\arguments{
\item{.Object}{an Authentication Object.}
}
\value{
A data.frame containing informoration about the X.509 certificate.
}
\description{
The DataONE X.509 certificate is read, if it is present and the
information contained in the certficate is returned as a data.frame.
}

