% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Client.R
\name{getCN}
\alias{getCN}
\alias{getCN,D1Client-method}
\title{Get the coordinating node associated with this D1Client object.}
\usage{
getCN(x)

\S4method{getCN}{D1Client}(x)
}
\arguments{
\item{x}{A D1Client object.}
}
\description{
Get the coordinating node associated with this D1Client object.
}
\note{
The method getCN has been deprecated.
}
\examples{
\dontrun{
cli <- D1Client("STAGING2", "urn:node:mnTestKNB")
testCN <- getCN(cli)
}
}
\seealso{
\code{\link[=D1Client-class]{D1Client}}{ class description.}
}
