\name{getD1Object}
\alias{getD1Object}
\title{Download a D1Object from the DataONE Cloud}
\arguments{
  \item{x}{: D1Client}

  \item{identifier}{: the identifier of the object to get}

  \item{...}{(not yet used)}
}
\value{
  the DataONE object
}
\description{
  Download a D1Object from the DataONE Cloud
}
\author{
  rnahf
}

