% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Object.R
\docType{methods}
\name{getIdentifier,D1Object-method}
\alias{getIdentifier,D1Object-method}
\title{Get the Identifier of the D1Object}
\usage{
\S4method{getIdentifier}{D1Object}(x)
}
\arguments{
\item{x}{D1Object}

\item{...}{(not yet used)}
}
\value{
the identifier
}
\description{
Get the Identifier of the D1Object
}
