\name{getIdentifiers}
\alias{getIdentifiers}
\title{Get the Identifiers of Package Members}
\arguments{
  \item{x}{: DataPackage}

  \item{...}{(not yet used)}
}
\value{
  list of identifiers
}
\description{
  Return the identifiers of the package members, as defined
  by the ResourceMap
}
\author{
  rnahf
}

