% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{methods}
\name{getMetadataMember}
\alias{getMetadataMember}
\alias{getMetadataMember,D1Client,DataPackage-method}
\title{Get the DataObject containing package metadata}
\usage{
getMetadataMember(x, dp, ...)

\S4method{getMetadataMember}{D1Client,DataPackage}(x, dp,
  as = "character", ...)
}
\arguments{
\item{x}{A D1Client object}

\item{dp}{A DataPackage object}

\item{...}{(Additional arguments, Not yet used.)}

\item{as}{A value of type \code{"character"} that specifies the return value. Possible values are \code{"character"} (the default) or \code{"DataPackage"}.}
}
\value{
The identifier of the metadata object
}
\description{
Each DataObject in the DataPackage is inspected to see if it matches one
of the formats supported by DataONE for metadata. If a package member's format matches
one of the supported formats, the identifier for that member is returned.
}
\details{
This method calls the DataONE CN 'format' service to obtain the current format
list.
}
