% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Client.R
\name{getMN}
\alias{getMN}
\alias{getMN,D1Client,ANY-method}
\alias{getMN,D1Client,character-method}
\title{Get a member node client based on its node identifier.}
\usage{
getMN(x, nodeid, ...)

\S4method{getMN}{D1Client,ANY}(x, nodeid, ...)

\S4method{getMN}{D1Client,character}(x, nodeid)
}
\arguments{
\item{x}{A D1Client object.}

\item{nodeid}{The identifier of the node to retrieve.}

\item{...}{(Not yet used)}
}
\description{
Get a member node client based on its node identifier.
}
\note{
This method has been superceded by \code{\link{getMNodeId}}
}
\examples{
\dontrun{
cli <- D1Client("STAGING2", "urn:node:mnTestKNB")
testMN <- getMN(cli)
}
}
\seealso{
\code{\link[=D1Client-class]{D1Client}}{ class description.}
}
