\name{getPackage}
\alias{getPackage}
\title{Download a DataPackage from the DataONE Cloud}
\arguments{
  \item{x}{D1Client}

  \item{identifier}{character: identifier of the
  ResourceMap}

  \item{...}{}
}
\value{
  the data package
}
\description{
  Given a valid identifier for a ResourceMap, instantiate
  an R DataPackage object containing all of the package
  members.
}
\author{
  rnahf
}

