% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\docType{methods}
\name{getTokenInfo}
\alias{getTokenInfo}
\alias{getTokenInfo,AuthenticationManager-method}
\title{Get authentication token information}
\usage{
getTokenInfo(.Object)

\S4method{getTokenInfo}{AuthenticationManager}(.Object)
}
\arguments{
\item{.Object}{an Authentication Object.}
}
\value{
A data.frame containing information about the authentication token.
}
\description{
The DataONE authentication token is read, if it has been
set, and the information it contains is returned as a data.frame.
}
