\name{helloJavaWorld}
\alias{helloJavaWorld}
\title{Function to say hello from a Java class}
\description{
  This function executes a method of a Java class
  defined in a JAR inside the R package.
}
\usage{
  helloJavaWorld()
}
\details{
  Details are given in the package vignette.
}
\value{
  The string "Hello Java World!" is returned.
}
\author{Tobias Verbeke}
\examples{
  helloJavaWorld()
}
\keyword{manip}