\name{insertRelationship}
\alias{insertRelationship}
\title{Associates Data Objects to the Science Metadata Objects that Describe Them}
\description{
  Associates Data Objects to the Science Metadata Objects
  that Describe Them
}
\note{
  Since the resource map that defines a package is separate
  from the items it associates, it is possible to use
  identifiers that have not been defined as members of the
  package.
}

