% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_request.R
\name{is_windows}
\alias{is_windows}
\title{Is the OS Windows?}
\usage{
is_windows()
}
\value{
TRUE if the OS is Windows, FALSE otherwise.
}
\description{
Check if the current operating system is Windows.
}
