% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\name{isAuthExpired}
\alias{isAuthExpired}
\alias{isAuthExpired,AuthenticationManager-method}
\title{Check if the currently valid authentication method has reached the expiration time.}
\usage{
isAuthExpired(.Object, ...)

\S4method{isAuthExpired}{AuthenticationManager}(.Object, node)
}
\arguments{
\item{.Object}{An AuthenticationManager instance}

\item{...}{(Not yet used)}

\item{node}{A D1Node instance}
}
\value{
A logical value: TRUE if authentication has expired, FALSE if not.
}
\description{
Check if the currently valid authentication method has reached the expiration time.
}
