% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CertificateManager.R
\name{isCertExpired}
\alias{isCertExpired}
\alias{isCertExpired,CertificateManager-method}
\title{Determine if an X.509 certificate has expired.}
\usage{
isCertExpired(x, ...)

\S4method{isCertExpired}{CertificateManager}(x)
}
\arguments{
\item{x}{a CertificateManager instance}

\item{...}{(Not yet used)}
}
\value{
TRUE if the certificate is expired
}
\description{
Returns \code{'TRUE'} if the certificate associated with a CertificateManager instance is expired.
A certificate is expired if any of the following conditions hold: 1) the current time is before or after the 
certificate validity dates, 2) the certificate is not valid according to a trusted Certificate Authority, or
3) no certificate can be found.
}
