% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNode.R
\docType{methods}
\name{listFormats}
\alias{listFormats}
\alias{listFormats,CNode-method}
\title{List all object formats registered in DataONE.}
\usage{
listFormats(x, ...)

\S4method{listFormats}{CNode}(x)
}
\arguments{
\item{x}{a valid CNode object}

\item{...}{(Not yet used)}
}
\value{
Returns a dataframe of all object formats registered in the DataONE Object Format Vocabulary.
}
\description{
The \link{listFormats} method queries a DataONE Coordinating Node for a 
list of all entries in the Object Format Vocabulary.
}
\examples{
library(dataone)
cn <- CNode()
formats <- listFormats(cn)
}
\seealso{
\code{\link[=CNode-class]{CNode}}{ class description.}
}

