% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{methods}
\name{listMemberNodes}
\alias{listMemberNodes}
\alias{listMemberNodes,D1Client-method}
\title{List DataONE Member Nodes.}
\usage{
listMemberNodes(x)

\S4method{listMemberNodes}{D1Client}(x)
}
\arguments{
\item{x}{A D1Client object.}
}
\description{
A D1Client object is associated with a DataONE Coordinating Node. The
\code{listMemberNodes} method lists all member nodes associated with a CN.
}
\examples{
{
d1c <- D1Client("PROD")
nodelist <- listMemberNodes(d1c)
}
}
\seealso{
\code{\link[=D1Client-class]{D1Client}}{ class description.}
}
