% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Node.R
\docType{methods}
\name{listQueryEngines}
\alias{listQueryEngines}
\alias{listQueryEngines,D1Node-method}
\title{Query a node for the list of query engines available on the node}
\usage{
listQueryEngines(x, ...)

\S4method{listQueryEngines}{D1Node}(x)
}
\arguments{
\item{x}{The CNode or MNode to list the query engines for.}

\item{...}{(Not yet used.)}
}
\value{
list The list of query engines.
}
\description{
Query a node for the list of query engines available on the node
}
\examples{
cn <- CNode("STAGING")
engines <- listQueryEngines(cn)
}

