% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\docType{methods}
\name{obscureAuth}
\alias{obscureAuth}
\alias{obscureAuth,AuthenticationManager-method}
\title{Temporarity disable DataONE authentication.}
\usage{
obscureAuth(.Object)

\S4method{obscureAuth}{AuthenticationManager}(.Object)
}
\arguments{
\item{.Object}{An AuthenticationManager instance}
}
\value{
The expiration date for the current authentication mechanism being used.
}
\description{
Calling \code{obscureAuth} temporarily disables authentication so that
}
\details{
This method is intended to be used for authentication testing.
\code{isAuthValid} will return FALSE. Authentication can be re-enabled by calling
\code{restoreAuth}.
}

