% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Node.R
\docType{methods}
\name{parseCapabilities}
\alias{parseCapabilities}
\alias{parseCapabilities,D1Node-method}
\title{Construct a Node, using a passed in capabilities XML}
\usage{
parseCapabilities(x, ...)

\S4method{parseCapabilities}{D1Node}(x, xml)
}
\arguments{
\item{x}{The node to which capabilities should be applied.}

\item{...}{(not yet used)}

\item{xml}{The XML capabilities representing the node to be created}
}
\value{
The Node object with modified capabilities properties from the XML
}
\description{
Construct a Node, using a passed in capabilities XML
}

