% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Node.R
\docType{methods}
\name{parseSolrResult}
\alias{parseSolrResult}
\alias{parseSolrResult,XMLInternalDocument-method}
\title{Parse Solr output into an R list}
\usage{
parseSolrResult(doc, ...)

\S4method{parseSolrResult}{XMLInternalDocument}(doc, parse, ...)
}
\arguments{
\item{doc}{The Solr result to parse, in XML format}

\item{...}{(Not yet used.)}

\item{parse}{A logical value, if TRUE the result is parsed to appropriate R types.}
}
\value{
resultList The Solr result as an R list
}
\description{
Solr output that is specified with a writer type of XML \code{'&wt="xml"'}
}
