% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Node.R
\name{ping}
\alias{ping}
\alias{ping,D1Node-method}
\title{Test if a node is online and accepting DataONE requests}
\usage{
ping(x, ...)

\S4method{ping}{D1Node}(x)
}
\arguments{
\item{x}{The CNode or MNode to check}

\item{...}{(Not yet used)}
}
\value{
logical A logical value set to TRUE if the node is up and FALSE if it is not

logical A logical value set to TRUE if the node is up and FALSE if it is not
}
\description{
Test if a node is online and accepting DataONE requests
}
\examples{
\dontrun{
cn <- CNode()
mn <- getMNode(cn, "urn:node:KNB")
isAlive <- ping(mn)
}
}
