\name{reserveIdentifier}
\alias{reserveIdentifier}
\title{Reserve an Identifier in the DataONE System}
\arguments{
  \item{x}{: D1Client}

  \item{id}{: identifier to reserve}

  \item{...}{(not yet used)}
}
\value{
  true if reserved
}
\description{
  Reserve an identifier for future use in the DataONE
  System.
}
\author{
  rnahf
}

