% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Client.R
\name{setMNodeId}
\alias{setMNodeId}
\alias{setMNodeId,D1Client,character-method}
\title{Set the member node identifier to be associated with the D1Client object.}
\usage{
setMNodeId(x, id)

\S4method{setMNodeId}{D1Client,character}(x, id)
}
\arguments{
\item{x}{A D1Client object.}

\item{id}{A DataONE member node identifier.}
}
\description{
The member node identifier is the URN identifier used by
DataONE to uniquely identifier a node, for example "urn:node:KNB" specifies
the "Knowledge Network for Biodiversity" member node.
}
\details{
One Member Node can be associated with the client as the default to which
data and metadata are written.
}
\seealso{
\code{\link[=D1Client-class]{D1Client}}{ class description.}
}
\author{
setMNodeId
}
