% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticationManager.R
\docType{methods}
\name{showAuth}
\alias{showAuth}
\alias{showAuth,AuthenticationManager-method}
\title{Display all authentication information}
\usage{
showAuth(.Object, ...)

\S4method{showAuth}{AuthenticationManager}(.Object, node)
}
\arguments{
\item{.Object}{An AuthenticationManager instance}

\item{...}{(Not yet used)}

\item{node}{A D1Node instance}
}
\description{
Display all authentication information
}
