% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{methods}
\name{convert.csv}
\alias{convert.csv}
\alias{convert.csv,D1Client-method}
\title{Convert a DataFrame to Standard CSV.}
\usage{
convert.csv(x, ...)

\S4method{convert.csv}{D1Client}(x, df, ...)
}
\arguments{
\item{x}{A D1Client object}

\item{...}{additional params passed to write.csv}

\item{df}{the dataFrame}
}
\value{
the dataframe serialized as a .csv
}
\description{
Convert a DataFrame to Standard CSV.
}
\examples{
\dontrun{ 
d1c <- D1Client("STAGING", "urn:node:mnStageUCSB2")
testdf <- data.frame(x=1:10,y=11:20)
sdf <- convert.csv(d1c, testdf)  
}
}
\seealso{
\code{\link[=D1Client-class]{D1Client}}{ class description.}
}
