% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Client.R
\docType{methods}
\name{D1Client}
\alias{D1Client}
\alias{D1Client,ANY,ANY-method}
\alias{D1Client,character,ANY-method}
\alias{D1Client,character,character-method}
\title{The DataONE client class used to downlaod, update and search for data in the DataONE network.}
\usage{
D1Client(x, y, ...)

\S4method{D1Client}{ANY,ANY}()

\S4method{D1Client}{character,ANY}(x, y, ...)

\S4method{D1Client}{character,character}(x, y)
}
\arguments{
\item{x}{The label for the DataONE environment to be using ('PROD','STAGING','SANDBOX','DEV')}

\item{y}{The node Id of the application's 'home' node.  Should be already registered to the corresponding 'env'}

\item{...}{(not yet used)}
}
\value{
the D1Client object representing the DataONE environment
}
\description{
The DataONE client class used to downlaod, update and search for data in the DataONE network.
}
\examples{
cli <- D1Client("PROD", "urn:node:KNB")
}
\seealso{
\code{\link[=D1Client-class]{D1Client}}{ class description.}
}

