% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Object.R
\docType{class}
\name{D1Object-class}
\alias{D1Object-class}
\title{D1Object (Deprecated) is a representation of a DataObject.}
\description{
D1Object has been deprecated in favor of datapack::DataObject, which provides
a wrapper for data and associated SystemMetadata.
}
\section{Slots}{

\describe{
\item{\code{dataObject}}{A backing instance of a DataObject, to which all methods and state are proxied}
}}

\section{Methods}{

\itemize{
 \item{\code{\link[dataone]{D1Object-initialize}}}{: Initialize a D1Object}
 \item{\code{\link[dataone]{getData}}}{: Get the data content of a specified D1Object.}
 \item{\code{\link[dataone]{getIdentifier}}}{: Get the identifier of the D1Object.}
 \item{\code{\link[dataone]{getFormatId}}}{: Get the formatId of the D1Object}
 \item{\code{\link[dataone]{setPublicAccess}}}{: Add a Rule to the AccessPolicy to make the object publicly readable.}
 \item{\code{\link[dataone]{canRead}}}{: Test whether the provided subject can read an object.}
 \item{\code{\link{asDataFrame}}}{: Return the D1Object as a data.frame.}
}
}

\seealso{
\code{\link{dataone}}{ package description.}
}
\keyword{classes}
