% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yyy.R
\name{dataone-defunct}
\alias{dataone-defunct}
\title{Defunct}
\description{
The following items are defunct in this release of dataone and are
no longer supported.
}
\section{These S4 methods are defunct}{

\itemize{
 \item{\code{\link{setObsoletedBy}}}{: Set a pid as being obsoleted by another pid.}
}
}

\keyword{internal}
