\name{getData}
\alias{getData}
\title{Get the Contents of the Specified Data Object}
\arguments{
  \item{x}{D1Object or DataPackage: the data structure from
  where to get the data}

  \item{id}{Missing or character: if}

  \item{...}{(not yet used)}
}
\value{
  character representation of the data
}
\description{
  Get the Contents of the Specified Data Object
}
\author{
  rnahf
}

