% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Object.R
\docType{methods}
\name{getData,D1Object-method}
\alias{getData,D1Object-method}
\title{Get the data content of a D1Object.}
\usage{
\S4method{getData}{D1Object}(x)
}
\arguments{
\item{x}{D1Object the data structure from where to get the data}
}
\description{
Get the data content of a D1Object.
}
