% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Node.R
\name{getErrorDescription}
\alias{getErrorDescription}
\title{Extract an error message from an http response.}
\usage{
getErrorDescription(response)
}
\arguments{
\item{response}{The httr response object to extract the error description from.}
}
\description{
Http requests can fail
for a variety of reasons, so getErrorDescription first tries to
determine what type of response was sent.
}
\details{
The return types handled by this function are:
  o An incorrect url is sent to DataONE and an error is returned by
    the web server, not a specified DataOne service url. In this case,
    a generic error message may be returned, e.g. status=404, URL not found
  o A DataOne service was called, and retunred an error message. In this
    case the DataONE response is parsed in an attemp to retrieve a
    meaningfull error message.
}

