% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Object.R
\name{getFormatId,D1Object-method}
\alias{getFormatId,D1Object-method}
\title{Get the FormatId of the D1Object}
\usage{
\S4method{getFormatId}{D1Object}(x)
}
\arguments{
\item{x}{D1Object}
}
\value{
the formatId
}
\description{
Get the FormatId of the D1Object
}
