% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Object.R
\name{setPublicAccess,D1Object-method}
\alias{setPublicAccess,D1Object-method}
\title{Make the object publicly readable.}
\usage{
\S4method{setPublicAccess}{D1Object}(x)
}
\arguments{
\item{x}{D1Object}
}
\value{
D1Object with modified access rules
}
\description{
This method should be called prior to creating the object in DataONE.  
When called before 
creating the object, adds a rule to the access policy that makes this object
publicly readable.  If called after creation, it will only change the system
metadata locally, and will not have any effect on remotely uploaded copies of
the D1Object.
}
\seealso{
\code{\link[=DataObject-class]{DataObject}}{ class description.}
}
